requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.1
beanie==1.29.0
cffi==1.17.1
click==8.1.8
cramjam==2.9.1
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-users==14.0.0
fastapi-users-db-beanie==3.0.0
fastparquet==2024.11.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
lazy-model==0.2.0
makefun==1.15.6
motor==3.6.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pwdlib==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.9.0
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2025.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
